# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: URL Diagnostics API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /status-pages:
    get:
      operationId: listStatusPages
      tags: [statusPages]
      summary: List status pages
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Status pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatusPage"
    post:
      operationId: createStatusPage
      tags: [statusPages]
      summary: Create a status page
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusPageInput"
      responses:
        "201":
          description: Status page created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusPage"
  /status-pages/{id}:
    put:
      operationId: updateStatusPage
      tags: [statusPages]
      summary: Update a status page
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusPageUpdate"
      responses:
        "200":
          description: Updated status page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusPage"
    delete:
      operationId: deleteStatusPage
      tags: [statusPages]
      summary: Delete a status page
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /public/status/{slug}:
    get:
      operationId: getPublicStatusPage
      tags: [statusPages]
      summary: Get public status page data (no auth)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public status page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicStatusPage"
        "404":
          description: Not found

  # ─── Tags ─────────────────────────────────────────────────────────────────
  /tags:
    get:
      operationId: listTags
      tags: [tags]
      summary: List all tags
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Tags list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
    post:
      operationId: createTag
      tags: [tags]
      summary: Create a tag
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
      responses:
        "201":
          description: Tag created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
  /tags/{id}:
    put:
      operationId: updateTag
      tags: [tags]
      summary: Update a tag
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagUpdate"
      responses:
        "200":
          description: Updated tag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
    delete:
      operationId: deleteTag
      tags: [tags]
      summary: Delete a tag
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
components:
  securitySchemes:
    clerkAuth:
      type: apiKey
      in: cookie
      name: __session
  schemas:
...
    Incident:
      type: object
      required: [id, monitorId, startedAt, isAcknowledged, createdAt]
      properties:
        id:
          type: string
        monitorId:
          type: string
        monitorName:
          type: ["string", "null"]
        monitorUrl:
          type: ["string", "null"]
        startedAt:
          type: string
          format: date-time
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        durationSeconds:
          type: ["integer", "null"]
        rootCause:
          type: ["string", "null"]
        affectedRegions:
          type: array
          items:
            type: string
        isAcknowledged:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    StatusPage:
      type: object
      required: [id, slug, title, isPublic, createdAt]
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        isPublic:
          type: boolean
        monitorIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    StatusPageInput:
      type: object
      required: [slug, title]
      properties:
        slug:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        description:
          type: string
        isPublic:
          type: boolean
          default: true
        monitorIds:
          type: array
          items:
            type: string
    StatusPageUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        isPublic:
          type: boolean
        monitorIds:
          type: array
          items:
            type: string
    PublicStatusPage:
      type: object
      required: [title, overallStatus, monitors]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        overallStatus:
          type: string
          enum: [operational, degraded, outage]
        monitors:
          type: array
          items:
            $ref: "#/components/schemas/PublicMonitorStatus"
        activeIncidents:
          type: array
          items:
            $ref: "#/components/schemas/Incident"
        recentIncidents:
          type: array
          items:
            $ref: "#/components/schemas/Incident"
    PublicMonitorStatus:
      type: object
      required: [name, status, uptime90d]
      properties:
        name:
          type: string
        status:
          type: string
          enum: [up, down, degraded, unknown]
        uptime90d:
          type: number
        dailyBars:
          type: array
          items:
            $ref: "#/components/schemas/DailyBar"
    DailyBar:
      type: object
      required: [date, status, uptime]
      properties:
        date:
          type: string
        status:
          type: string
          enum: [up, down, degraded, nodata]
        uptime:
          type: number

    # ─── Tags ─────────────────────────────────────────────────────────────
    Tag:
      type: object
      required: [id, name, color]
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
    TagInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
          default: "#6366f1"
    TagUpdate:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
...