# lib/api-zod/src/generated/types/dailyBar.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * URL Diagnostics API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DailyBarStatus } from './dailyBarStatus';

export interface DailyBar {
  date: string;
  status: DailyBarStatus;
  uptime: number;
}